requirements.txt
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.5.2
astroid==2.11.5
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
black==21.4b2
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coverage==5.5
cryptography==37.0.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
Django==3.1.10
django-allauth==0.44.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.11.2
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
django-model-utils==4.1.1
django-redis==4.12.1
django-stubs==1.8.0
django-stubs-ext==0.4.0
docutils==0.16
executing==0.8.3
factory-boy==3.2.0
Faker==13.11.1
filelock==3.7.0
flake8==3.9.1
flake8-isort==4.0.0
identify==2.5.0
idna==3.3
imagesize==1.3.0
iniconfig==1.1.1
ipdb==0.13.7
ipython==8.3.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
lazy-object-proxy==1.7.1
livereload==2.6.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.5.2
pluggy==0.13.1
pre-commit==2.12.1
prompt-toolkit==3.0.29
psycopg2-binary==2.8.6
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
Pygments==2.12.0
PyJWT==2.4.0
pylint==2.13.9
pylint-django==2.4.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pytest==6.2.3
pytest-django==4.2.0
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
qrcode==7.3.1
redis==3.5.3
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==3.5.4
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stack-data==0.2.0
termcolor==1.1.0
testfixtures==6.18.5
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.2.1.post0
typed-ast==1.4.3
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.14.1
